requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp_socks==0.10.1
aiosignal==1.3.2
attrs==25.1.0
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
python-socks==2.7.1
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.3.0
wsproto==1.2.0
yarl==1.18.3
